fix(ci): resolve rust-security failures
- Cargo.toml: GPL-3.0 → GPL-3.0-only (deprecated SPDX identifier)
- deny.toml: ignore RUSTSEC-2023-0071 (rsa Marvin timing, no fix
  available upstream; pulled in by sqlx-mysql, which we don't use)
- Cargo.lock: bump aws-lc-rs 1.15→1.16, aws-lc-sys 0.37→0.40
  (fixes RUSTSEC-2026-0044/0045/0046/0047/0048), quinn-proto 0.11.13
  →0.11.14 (fixes RUSTSEC-2026-0037), rustls-webpki 0.103.9→0.103.13
  (fixes RUSTSEC-2026-0104)
- auth_test.rs: remove useless .into_iter() on get_all() (clippy)
- concurrency_test.rs: remove redundant as i32 casts (clippy)

+++ apps/server/tests/integration/concurrency_test.rs
@@ -219,7 +219,7 @@ async fn test_concurrent_same_errors_same_project_groups_into_one_issue() {
 
     assert_eq!(issues.len(), 1, "Expected 1 issue, got {}", issues.len());
     assert_eq!(
-        issues[0].digested_event_count, num_events as i32,
+        issues[0].digested_event_count, num_events,
         "Expected {} events in issue, got {}",
         num_events, issues[0].digested_event_count
     );
@@ -513,7 +513,7 @@ async fn test_concurrent_mixed_create_and_update() {
     // Verify each issue has 4 events
     for issue in &issues {
         assert_eq!(
-            issue.digested_event_count, copies_per_error as i32,
+            issue.digested_event_count, copies_per_error,
             "Each issue should have {} events, issue {} has {}",
             copies_per_error, issue.id, issue.digested_event_count
         );
